increased max level 2 day to 180

+++ bbn/src/org/cougaar/logistics/plugin/trans/base/GLMTransOneToManyExpanderPlugin.java
@@ -89,7 +89,7 @@ public class GLMTransOneToManyExpanderPlugin extends UTILExpanderPluginAdapter i
   protected OperatingMode level2Horizon, level6Horizon;
 
   public final Integer LEVEL_2_MIN = new Integer(40); // later, these should be parameters to plugin...
-  public final Integer LEVEL_2_MAX = new Integer(80);
+  public final Integer LEVEL_2_MAX = new Integer(180);
   public final Integer LEVEL_6_MIN = new Integer(20);
   public final Integer LEVEL_6_MAX = new Integer(180);
   public final int LEVEL_6_MODE = 0;
